+++ docs/changelog.md
@@ -56,6 +56,10 @@
 
 ## 更新说明
 
+1.9.78 fix: fix macro_usa_gdp_monthly interface
+
+    1. 修复 macro_usa_gdp_monthly 接口
+
 1.9.77 fix: fix stock_hot_rank_wc interface
 
     1. 修复 stock_hot_rank_wc 接口
@@ -2286,6 +2290,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.9.78 fix: fix macro_usa_gdp_monthly interface
 
 1.9.77 fix: fix stock_hot_rank_wc interface